refactor: move votingNeuronSelectStore to voting
